travel_backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.5.0
psycopg2-binary==2.9.10
python-decouple==3.8
langchain==0.3.27
langchain-openai==0.3.28
langchain-community==0.3.27
langchain-core==0.3.73
openai==1.99.2
google-search-results==2.4.2
requests==2.32.4
python-dotenv==1.1.1
